Cleaned up code styling a bit
